requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
pypdf
google-genai
motor
dnspython
pydantic-settings
certifi
